requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-Cors==3.0.10
requests==2.31.0
python-dotenv==1.0.0
zhipuai==2.1.5.20241204
google-cloud-vision
google-cloud-translate
backoff
gunicorn